+++ src/components/Flash.d.ts
