+++ src/main.tsx
@@ -4,7 +4,7 @@ import "@fontsource-variable/inter";
 import "@fontsource-variable/jetbrains-mono";
 
 import "./index.css";
-import App from "./app/App";
+import App from "./App";
 
 const rootElement = document.getElementById("root");
 if (!rootElement) throw new Error("Failed to find the root element");
